Refactoring to check loss weighting

+++ src/classes/ClassConditionalLSTMPostMeanDiffTrainer.py
@@ -43,7 +43,7 @@ class ConditionalLSTMPostMeanDiffusionModelTrainer(nn.Module):
         self.loss_fn = loss_fn  # If callable, need to ensure we allow for gradient computation
         self.loss_aggregator = loss_aggregator().to(self.device_id)
         self.loss_factor = loss_factor
-        self.deltaT = torch.Tensor([deltaT], dtype=torch.float32).to(self.device_id)
+        self.deltaT = torch.Tensor([deltaT]).to(self.device_id)
 
         self.diffusion = diffusion
         self.train_eps = train_eps
