+++ src/classes/ClassConditionalLSTMWithPositionDiffTrainer.py
@@ -42,7 +42,7 @@ class ConditionalLSTMWithPositionDiffusionModelTrainer(nn.Module):
         self.train_loader = train_data_loader
         self.loss_fn = loss_fn  # If callable, need to ensure we allow for gradient computation
         self.loss_aggregator = loss_aggregator().to(self.device_id)
-        self.deltaT = torch.Tensor([deltaT], dtype=torch.float32).to(self.device_id)
+        self.deltaT = torch.Tensor([deltaT]).to(self.device_id)
 
         self.diffusion = diffusion
         self.train_eps = train_eps
