+++ src/classes/ClassConditionalMarkovianWithPositionDiffTrainer.py
@@ -54,7 +54,7 @@ class ConditionalMarkovianWithPositionDiffusionModelTrainer(nn.Module):
         self.train_loader = train_data_loader
         self.loss_fn = loss_fn  # If callable, need to ensure we allow for gradient computation
         self.loss_aggregator = loss_aggregator().to(self.device_id)
-        self.deltaT = torch.Tensor([deltaT], dtype=torch.float32).to(self.device_id)
+        self.deltaT = torch.Tensor([deltaT]).to(self.device_id)
 
         self.diffusion = diffusion
         self.train_eps = train_eps
